refactor: move the permissions to their own package
Just cleaning up the project structure.

RHCLOUD-34485
