+++ backend/src/main/java/com/redhat/cloud/notifications/auth/kessel/permission/WorkspacePermission.java
@@ -1,4 +1,4 @@
-package com.redhat.cloud.notifications.auth.kessel;
+package com.redhat.cloud.notifications.auth.kessel.permission;
 
 public enum WorkspacePermission implements KesselPermission {
     CREATE_DRAWER_INTEGRATION("notifications_integration_subscribe_drawer"),
